Address review feedback and fix tests
