+++ subprojects/dependency-management/src/main/java/org/gradle/api/internal/artifacts/transform/ArtifactTransformDependenciesProvider.java
@@ -17,26 +17,8 @@
 package org.gradle.api.internal.artifacts.transform;
 
 import org.gradle.api.artifacts.transform.ArtifactTransformDependencies;
-import org.gradle.api.file.FileCollection;
 import org.gradle.api.internal.attributes.ImmutableAttributes;
-import org.gradle.api.internal.file.collections.ImmutableFileCollection;
 
 interface ArtifactTransformDependenciesProvider {
-
-    ArtifactTransformDependencies EMPTY_DEPENDENCIES = new ArtifactTransformDependencies() {
-        @Override
-        public FileCollection getFiles() {
-            return ImmutableFileCollection.of();
-        }
-    };
-
-    ArtifactTransformDependenciesProvider EMPTY = new ArtifactTransformDependenciesProvider() {
-
-        @Override
-        public ArtifactTransformDependencies forAttributes(ImmutableAttributes attributes) {
-            return EMPTY_DEPENDENCIES;
-        }
-    };
-
     ArtifactTransformDependencies forAttributes(ImmutableAttributes attributes);
 }
