+++ subprojects/dependency-management/src/main/java/org/gradle/api/internal/artifacts/transform/DefaultTransformerInvoker.java
@@ -96,7 +96,7 @@ public class DefaultTransformerInvoker implements TransformerInvoker {
 
     @Override
     public boolean hasCachedResult(Transformer transformer, File primaryInput, ArtifactTransformDependencies dependencies) {
-        return historyRepository.hasCachedResult(getImmutableTransformationIdentity(primaryInput, transformer ,dependencies));
+        return historyRepository.hasCachedResult(getImmutableTransformationIdentity(primaryInput, transformer, dependencies));
     }
 
     @Override
