+++ subprojects/dependency-management/src/test/groovy/org/gradle/api/internal/artifacts/transform/DefaultVariantTransformRegistryTest.groovy
@@ -80,7 +80,7 @@ class DefaultVariantTransformRegistryTest extends Specification {
         !outputFile.exists()
 
         when:
-        def transformed = registration.transformationStep.transform(TransformationSubject.initial(TEST_INPUT), ArtifactTransformDependenciesProvider.EMPTY).files
+        def transformed = registration.transformationStep.transform(TransformationSubject.initial(TEST_INPUT), DefaultArtifactTransformDependenciesProvider.EMPTY).files
 
         then:
         transformed.size() == 1
@@ -119,7 +119,7 @@ class DefaultVariantTransformRegistryTest extends Specification {
         !outputFile.exists()
 
         when:
-        def transformed = registration.transformationStep.transform(TransformationSubject.initial(TEST_INPUT), ArtifactTransformDependenciesProvider.EMPTY).files
+        def transformed = registration.transformationStep.transform(TransformationSubject.initial(TEST_INPUT), DefaultArtifactTransformDependenciesProvider.EMPTY).files
 
         then:
         transformed.collect { it.name } == ['OUTPUT_FILE', 'EXTRA_1', 'EXTRA_2']
@@ -154,7 +154,7 @@ class DefaultVariantTransformRegistryTest extends Specification {
 
         when:
         def registration = registry.transforms.first()
-        def result = registration.transformationStep.transform(TransformationSubject.initial(TEST_INPUT), ArtifactTransformDependenciesProvider.EMPTY)
+        def result = registration.transformationStep.transform(TransformationSubject.initial(TEST_INPUT), DefaultArtifactTransformDependenciesProvider.EMPTY)
 
         then:
         def failure = result.failure
@@ -191,7 +191,7 @@ class DefaultVariantTransformRegistryTest extends Specification {
 
         when:
         def registration = registry.transforms.first()
-        def failure = registration.transformationStep.transform(TransformationSubject.initial(TEST_INPUT), ArtifactTransformDependenciesProvider.EMPTY).failure
+        def failure = registration.transformationStep.transform(TransformationSubject.initial(TEST_INPUT), DefaultArtifactTransformDependenciesProvider.EMPTY).failure
 
         then:
         failure instanceof ObjectInstantiationException
@@ -225,7 +225,7 @@ class DefaultVariantTransformRegistryTest extends Specification {
 
         when:
         def registration = registry.transforms.first()
-        def failure = registration.transformationStep.transform(TransformationSubject.initial(TEST_INPUT), ArtifactTransformDependenciesProvider.EMPTY).failure
+        def failure = registration.transformationStep.transform(TransformationSubject.initial(TEST_INPUT), DefaultArtifactTransformDependenciesProvider.EMPTY).failure
 
         then:
         failure.message == 'broken'
@@ -320,7 +320,7 @@ class DefaultVariantTransformRegistryTest extends Specification {
     }
 
     private void runTransformer(File input) {
-        1 * transformerInvoker.invoke(_ as Transformer, input, _ as TransformationSubject, _ as ArtifactTransformDependencies)  >> { Transformer transformer, File primaryInput, TransformationSubject subject, ArtifactTransformDependencies dependencies ->
+        1 * transformerInvoker.invoke(_ as Transformer, input, _ as ArtifactTransformDependencies, _ as TransformationSubject)  >> { Transformer transformer, File primaryInput, ArtifactTransformDependencies dependencies, TransformationSubject subject ->
             return Try.ofFailable { ImmutableList.copyOf(transformer.transform(primaryInput, outputDirectory, dependencies)) }
         }
     }
